refactor: components name and paths
